pyproject.toml
[project]
name = "rhiza-education"
version = "0.1.0"
description = "Educational material for Rhiza — the living template system."
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "PyYAML>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
